Remote authentication proof of concept, unify protobuf code with server
