+++ .gitignore
@@ -1,5 +1,6 @@
 target
 generated-src
+.recommenders
 .classpath
 .metadata
 .project
