--- authenticationTest/README.md
@@ -1,8 +1,0 @@
-To build and run this project, you should do the following:
-
-Make sure you have the protobuf compiler (protoc) and maven installed.
-You can download the protoc binaries at: https://github.com/google/protobuf/releases
-
-Compile the protobuf files, and run the service using mvn:
-
-$ mvn spring-boot:run
