+++ server/build-proto.sh
@@ -1,0 +1,3 @@
+mkdir generated-src -p
+rm -rf generated-src/*
+protoc -I protobuf --java_out=generated-src protobuf/*.proto
