+++ server/pom.xml
@@ -43,6 +43,16 @@
       <artifactId>jquery</artifactId>
       <version>3.1.0</version>
     </dependency>
+    <dependency>
+      <groupId>com.google.protobuf</groupId>
+      <artifactId>protobuf-java</artifactId>
+      <version>3.1.0</version>
+    </dependency>
+    <dependency>
+      <groupId>com.google.guava</groupId>
+      <artifactId>guava</artifactId>
+      <version>19.0</version>
+    </dependency>
 
     <!-- <dependency>
       <groupId>org.springframework.boot</groupId>
