+++ server/protobuf/tinyfs_client_credentials.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package tinyfs;
 
-option java_package = "com.avilan.tinyfsservice";
+option java_package = "com.tinyfs.auth";
 option java_outer_classname = "ClientCredentialsProto";
 
 message AccessCredentials {
