+++ server/src/main/java/com/tinyfs/config/WebSocketConfig.java
@@ -1,12 +1,13 @@
-package foo;
+package com.tinyfs.config;
 
-import foo.FooHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.WebSocketHandler;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
+
+import com.tinyfs.handler.FooHandler;
 
 @Configuration
 @EnableWebSocket
