+++ server/src/main/java/com/tinyfs/exception/CredentialsException.java
@@ -1,0 +1,11 @@
+package com.tinyfs.exception;
+
+public class CredentialsException extends Exception {
+
+  private static final long serialVersionUID = 1L;
+
+  public CredentialsException(Exception e) {
+    super(e);
+  }
+
+}
